% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aln}
\alias{aln}
\title{\emph{Rhodopirellula} sp. alignemnts used for Zure et al., 2016}
\format{A DNAbin object with 361890 sequences of 483 nucleotides each}
\usage{
aln
}
\description{
A dataset containing the aligned reads for the study described in 
 Zure et al., 2016. The variables are as follows:
}
\references{
Zure, M., Fernandez-Guerra, A., Munn C. and Harder, J. (2016). High resolution distribution of 
 closely related \emph{Rhodopirellula} species in European coastal sediments. Submitted.
}
\keyword{datasets}

