% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entropy}
\alias{entropy}
\title{Entropy values for each position of the alignment used for Zure et al., 2016}
\format{A data frame with 483 rows and 2 variables:
\itemize{
  \item V1: nucleotide positions of the alignment
  \item V2: entropy value
}}
\usage{
entropy
}
\description{
A dataset containing the entropy values for each 483 nucleotides contained in the 
 alignemnt. This is a raw dataframe containing the results of running \href{https://github.com/meren/oligotyping}{entropy-analysis} 
command from oligotyping package.
}
\references{
Eren, A. M., Morrison, H. G., Lescault, P. J., Reveillaud, J., Vineis, J. H., & Sogin, M. L. (2015). 
Minimum entropy decomposition: Unsupervised oligotyping for sensitive partitioning of high-throughput marker gene 
sequences. The ISME Journal, 9(4), 968-979

Zure, M., Fernandez-Guerra, A., Munn C. and Harder, J. (2016). High resolution distribution of 
 closely related \emph{Rhodopirellula} species in European coastal sediments. Submitted.
}
\keyword{datasets}

