% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_3rd_codon.R
\name{get_3rd_codon}
\alias{get_3rd_codon}
\title{Create a sequence alignment with only the 3rd codon}
\usage{
get_3rd_codon(aln, ...)
}
\arguments{
\item{aln}{a sequence alignment}
}
\description{
Create a sequence alignment with only the 3rd codon
}

