% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oligo4fun.R
\docType{package}
\name{oligo4fun}
\alias{oligo4fun}
\alias{oligo4fun-package}
\title{Minimum Entropy Decomposition diagnostics for functional genes}
\description{
oligo4fun lets you diagnose if you should use all codon 
positions or remove the 3rd position before oligotyping analysis.
}
\author{
Antonio Fernandez-Guerra, \email{afernand@mpi-bremen.de}

Marina Zure, \email{mzure@mpi-bremen.de}
}
\references{
\url{https://github.com/genomewalker/oligo4fun}
}

