% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_entropy_density.R
\name{plot_entropy_density}
\alias{plot_entropy_density}
\title{Plot the distribution of entropy values}
\usage{
plot_entropy_density(data, process = TRUE)
}
\arguments{
\item{data}{a dataframe containing the results of running \href{https://github.com/meren/oligotyping}{entropy-analysis} 
command from oligotyping package. It could be a raw file obtained after the entropy-analysis or a processed raw file by
\code{\link{process_entropy_file}}}

\item{process}{a logical indicating whether to process the entropy data; defaults to TRUE so}
}
\value{
a ggplot2 object with
}
\description{
Plot the distribution of entropy values
}
\examples{
entropy_plot <- plot_entropy_density(entropy)
}
\references{
Eren, A. M., Morrison, H. G., Lescault, P. J., Reveillaud, J., Vineis, J. H., & Sogin, M. L. (2015). 
Minimum entropy decomposition: Unsupervised oligotyping for sensitive partitioning of high-throughput marker gene 
sequences. The ISME Journal, 9(4), 968-979
}

