% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_saturation.R
\name{plot_saturation}
\alias{plot_saturation}
\title{Draw a saturation plot of the oligotyping input file}
\usage{
plot_saturation(aln = aln, nseqs = 1000, model = "K80", all = FALSE,
  verbose = TRUE, seed = 0, rsamp = FALSE, ...)
}
\arguments{
\item{aln}{a matrix containing the DNA sequences; this must be of class "DNAbin"}

\item{nseqs}{a value for the number of sequences to pick}

\item{model}{a character string specifying the evolutionary model to be used by \code{\link[ape]{dist.dna}}}

\item{all}{a logical indicating whether to use all codon positions; defaults to FALSE so only the third codon position is used.}

\item{verbose}{a logical indicating whether to show in screen the progress; defaults to TRUE.}

\item{seed}{a value for the random number generator. This will allow you to repeat the analysis}
}
\value{
An object of class \dQuote{oligodiag} is a list containing at least the following components:  
\describe{
 \item{plot}{a ggplot object containing the saturation plot}
 \item{seed}{the seed used for picking the random sequences}
 \item{aln}{a matrix of the random selected sequences stored in binary format}
 \item{stats}{mean and standard deviation of transitions and transversions}
 \item{all_codons}{logical if all codon positions have been used}
 \item{saturation}{whether your alignemnt possible presents saturation}
}
}
\description{
Draw a saturation plot of the oligotyping input file
}
\details{
Usually oligotyping datasets are very large and perform a full analysis would be computationally really expensive. We
recommend to use up to 5000 random sequences in a desktop computer. You can calculate multiple saturation plots and its associated 
statistics using \code{\link[oligo4fun]{plot_saturation_n}} to get a better overview of your alignment.
}
\examples{
saturation_plot <- plot_saturation(aln, nseqs = 1000, all = FALSE)
}

