% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_saturation_n.R
\name{plot_saturation_n}
\alias{plot_saturation_n}
\title{Draw a saturation plot of the oligotyping input file (n) times}
\usage{
plot_saturation_n(aln = aln, rep = 100, nseqs = 1000, model = "K80",
  parallel = FALSE, all = FALSE, reg_id = NULL, reg_dir = NULL,
  conf_file = NULL, job_res = list(), ...)
}
\arguments{
\item{aln}{a matrix containing the DNA sequences; this must be of class "DNAbin"}

\item{rep}{number of random subsamplings}

\item{nseqs}{a value for the number of sequences to pick}

\item{model}{a character string specifying the evolutionary model to be used by \code{\link[ape]{dist.dna}}}

\item{parallel}{a logical indicating whether to do the random subsampling on a sequential way or use the \code{\href{https://github.com/tudo-r/BatchJobs}{BatchJobs}}
framework to distribute the random subsamplings on different cores or in computer cluster}

\item{all}{a logical indicating whether to use all codon positions; defaults to FALSE so only the third codon position is used.}

\item{reg_id}{Name of registry. Displayed e.g. in mails or in cluster queue}

\item{reg_dir}{Path where files regarding the registry / jobs should be saved}

\item{conf_file}{Location of the configuration file to load}
}
\value{
An object of class \dQuote{oligodiag} is a list containing at least the following components:  
\describe{
 \item{plot}{a ggplot object containing the saturation plots of each random sampling}
 \item{seed}{the seeds used for picking the random sequences}
 \item{aln}{a matrix of the random selected sequences stored in binary format for each random subsampling}
 \item{combined_stats}{mean and standard deviation of transitions and transversions for each random subsampling}
 \item{saturation}{whether your alignment presents saturation for each random subsampling}
 \item{raw}{raw results for each random subsampling}
}
}
\description{
Perform multiple random subsamplings of your alignemnt.
}
\details{
You can calculate multiple saturation plots and its associated statistics. The different
random subsamplings can be easily distributed over multiple cores or in a computer cluster.
}
\examples{
saturation_plots <- plot_saturation_n(aln, nseqs = 1000, rep = 100, parallel = F)
}

