% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_entropy_file.R
\name{process_entropy_file}
\alias{process_entropy_file}
\title{Pre-process entropy-analysis file}
\usage{
process_entropy_file(file = file)
}
\arguments{
\item{file}{a dataframe containing the results after running 
\href{https://github.com/meren/oligotyping}{entropy-analysis} command from oligotyping package}
}
\value{
A dataframe with the nucleotide positions, the entropy values and the codon position
}
\description{
Pre-process the entropy-analysis file containing the results after running 
\href{https://github.com/meren/oligotyping}{entropy-analysis} command from oligotyping package
}
\details{
Transforms the generated entropy file for plotting.
}
\examples{
entropy_file <- process_entropy_file(entropy)
}
\references{
Eren, A. M., Morrison, H. G., Lescault, P. J., Reveillaud, J., Vineis, J. H., & Sogin, M. L. (2015). 
Minimum entropy decomposition: Unsupervised oligotyping for sensitive partitioning of high-throughput marker gene 
sequences. The ISME Journal, 9(4), 968-979
}

