% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sequences.R
\name{random_sequences}
\alias{random_sequences}
\title{Pick random sequences from the alignment}
\usage{
random_sequences(aln = aln, nseqs = 1000, seed = 0, verbose = verbose,
  ...)
}
\arguments{
\item{aln}{a sequence alignment}

\item{nseqs}{number of sequences to pick}

\item{seed}{choose a seed for the random number generator. This will allow you to repeat the analysis.}
}
\value{
An object of class \dQuote{oligodiag} is a list containing at least the following components:  
\describe{
 \item{seed}{the seed used for picking the random sequences}
 \item{aln}{a matrix of the random selected sequences stored in binary format}
}
}
\description{
Pick random sequences from the alignment
}
\examples{
rand_seqs <- random_sequences(aln)
}

