% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_3rd_codon.R
\name{remove_3rd_codon}
\alias{remove_3rd_codon}
\title{Create a sequence alignment with only the 1st and 2nd codon}
\usage{
remove_3rd_codon(aln, ...)
}
\arguments{
\item{aln}{a sequence alignment}
}
\description{
Create a sequence alignment with only the 1st and 2nd codon
}

