% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_saturation_alignments.R
\name{write_saturation_alignments}
\alias{write_saturation_alignments}
\title{Write alignments from each saturation test random sampling to disk}
\usage{
write_saturation_alignments(data, dir = NULL, no3rd = FALSE, ...)
}
\arguments{
\item{data}{a \dQuote{oligodiag} object}

\item{dir}{path where the alignment should be saved}

\item{no3rd}{a logical wether to print the alignment with all codon position or the}
}
\description{
Write alignments from each saturation test random sampling to disk
}
\examples{
write_saturation_alignments(data)
}

