% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_saturation_plots.R
\name{write_saturation_plots}
\alias{write_saturation_plots}
\title{Write saturation plots from each saturation test random sampling to disk}
\usage{
write_saturation_plots(data, dir = NULL, format = "png", ...)
}
\arguments{
\item{data}{a \dQuote{oligodiag} object}

\item{dir}{path where the plot should be saved}

\item{format}{image format for the saved plot. Should be \dQuote{png}, \dQuote{jpg} or \dQuote{pdf}}
}
\description{
Write saturation plots from each saturation test random sampling to disk
}
\examples{
write_saturation_plots(data)
}

